Extract GDAL dependency from libsirius/libsirius-static
* create `libsirius-gdal-static` target (used in tests and sirius main)
* add a CMake option to disable `sirius` executable target from default build
* update FFTW3 find_package
* update sirius install directory content
