+++ CMakeLists.txt
@@ -27,7 +27,8 @@ include(cmake/BuildType.cmake)
 set(SIRIUS_VERSION "0.0.0" CACHE STRING "Sirius version")
 set(SIRIUS_REVISION_COMMIT "sirius-no-revision-commit" CACHE STRING "Sirius revision commit")
 
-option(ENABLE_CACHE_OPTIMIZATION "Enable cache (FFTW plan, Filter FFT)" ON)
+option(ENABLE_SIRIUS_EXECUTABLE "Enable Sirius executable target" ON)
+option(ENABLE_CACHE_OPTIMIZATION "Enable cache optimization (FFTW plan, Filter FFT)" ON)
 option(ENABLE_LOGS "Enable logs" ON)
 option(ENABLE_GSL_CONTRACTS "Enable GSL contracts" OFF)
 option(ENABLE_DOCUMENTATION "Enable documentation generation" OFF)
@@ -71,6 +72,7 @@ endif()
 message(STATUS "Sirius ${SIRIUS_VERSION} (${SIRIUS_REVISION_COMMIT})")
 message(STATUS "Build type: ${CMAKE_BUILD_TYPE}")
 message(STATUS "Install directory: ${CMAKE_INSTALL_PREFIX}")
+message(STATUS "Enable Sirius executable: ${ENABLE_SIRIUS_EXECUTABLE}")
 message(STATUS "Enable cache: ${ENABLE_CACHE_OPTIMIZATION}")
 message(STATUS "Enable logs: ${ENABLE_LOGS}")
 message(STATUS "Enable GSL contracts: ${ENABLE_GSL_CONTRACTS}")
