+++ cmake/FindFFTW3.cmake
@@ -33,12 +33,27 @@
 #   FFTW3_LIBRARIES    - List of libraries when using fftw3.
 #   FFTW3_FOUND        - True if fftw3 found.
 
+# this implementation is inspired by
+#    https://github.com/InsightSoftwareConsortium/ITK/blob/master/CMake/FindFFTW.cmake
+
 find_package(PkgConfig)
 
 pkg_check_modules(PC_FFTW3 fftw3)
-find_path(FFTW3_INCLUDE_DIR fftw3.h HINTS ${PC_FFTW3_INCLUDE_DIRS})
 
-find_library(FFTW3_LIBRARY NAMES fftw3 HINTS ${PC_FFTW3_LIBRARY_DIRS} )
+set(FFTW_LIB_SEARCHPATH
+    ${PC_FFTW3_INCLUDE_DIRS}
+    /usr/lib/fftw
+    /usr/local/lib/fftw)
+
+if (${FFTW_INCLUDE_PATH})
+    get_filename_component(FFTW_INSTALL_BASE_PATH ${FFTW_INCLUDE_PATH} PATH)
+    list(APPEND FFTW_LIB_SEARCHPATH
+        ${FFTW_INSTALL_BASE_PATH}/lib
+        ${FFTW_INSTALL_BASE_PATH}/lib64)
+endif()
+
+find_path(FFTW3_INCLUDE_DIR fftw3.h HINTS ${FFTW_LIB_SEARCHPATH})
+find_library(FFTW3_LIBRARY NAMES fftw3 HINTS ${FFTW_LIB_SEARCHPATH} )
 
 set(FFTW3_LIBRARIES ${FFTW3_LIBRARY} )
 set(FFTW3_INCLUDE_DIRS ${FFTW3_INCLUDE_DIR} )
