+++ src/CMakeLists.txt
@@ -21,7 +21,7 @@
 
 cmake_minimum_required(VERSION 3.2)
 
-set(SIRIUS_SRC
+set(LIBSIRIUS_SRC
     sirius/exception.h
 
     sirius/types.h
@@ -35,9 +35,6 @@ set(SIRIUS_SRC
     sirius/i_frequency_resampler.h
     sirius/frequency_resampler_factory.h
     sirius/frequency_resampler_factory.cc
-
-    sirius/image_streamer.h
-    sirius/image_streamer.cc
 
     # resampler
     sirius/resampler/frequency_resampler.h
@@ -64,27 +61,11 @@ set(SIRIUS_SRC
     sirius/fftw/wrapper.h
     sirius/fftw/wrapper.cc
 
-    # gdal
-    sirius/gdal/error_code.h
-    sirius/gdal/error_code.cc
-    sirius/gdal/exception.h
-    sirius/gdal/exception.cc
-    sirius/gdal/stream_block.h
-    sirius/gdal/input_stream.h
-    sirius/gdal/input_stream.cc
-    sirius/gdal/resampled_output_stream.h
-    sirius/gdal/resampled_output_stream.cc
-    sirius/gdal/types.h
-    sirius/gdal/wrapper.h
-    sirius/gdal/wrapper.cc
-
     # utils
     sirius/utils/concurrent_queue.h
     sirius/utils/concurrent_queue.txx
     sirius/utils/concurrent_queue_error_code.h
     sirius/utils/concurrent_queue_error_code.cc
-    sirius/utils/debug.h
-    sirius/utils/debug.cc
     sirius/utils/gsl.h
     sirius/utils/log.h
     sirius/utils/log.cc
@@ -99,14 +80,14 @@ set(SIRIUS_CONFIG_OUT ${CMAKE_CURRENT_BINARY_DIR}/include/sirius/sirius.h)
 
 configure_file(${SIRIUS_CONFIG_IN} ${SIRIUS_CONFIG_OUT} @ONLY)
 
-list(APPEND SIRIUS_INCLUDE_DIRS
+list(APPEND LIBSIRIUS_INCLUDE_DIRS
         ${CMAKE_CURRENT_SOURCE_DIR}
         ${CMAKE_CURRENT_BINARY_DIR}/include)
-LIST(APPEND SIRIUS_LINK_LIBS "gdal" "fftw3" "spdlog" "gsl")
+LIST(APPEND LIBSIRIUS_LINK_LIBS "fftw3" "spdlog" "gsl")
 
-add_library(libsirius SHARED ${SIRIUS_SRC})
+add_library(libsirius SHARED ${LIBSIRIUS_SRC})
 set_property(TARGET libsirius PROPERTY POSITION_INDEPENDENT_CODE ON)
-add_library(libsirius-static STATIC ${SIRIUS_SRC})
+add_library(libsirius-static STATIC ${LIBSIRIUS_SRC})
 set_property(TARGET libsirius-static PROPERTY POSITION_INDEPENDENT_CODE ON)
 
 if (${ENABLE_LOGS})
@@ -121,26 +102,62 @@ if (${ENABLE_CACHE_OPTIMIZATION})
     target_compile_definitions(libsirius-static PUBLIC SIRIUS_ENABLE_CACHE_OPTIMIZATION=1)
 endif ()
 
-target_include_directories(libsirius PUBLIC ${SIRIUS_INCLUDE_DIRS})
-target_include_directories(libsirius-static PUBLIC ${SIRIUS_INCLUDE_DIRS})
+target_include_directories(libsirius PUBLIC ${LIBSIRIUS_INCLUDE_DIRS})
+target_include_directories(libsirius-static PUBLIC ${LIBSIRIUS_INCLUDE_DIRS})
 
-target_link_libraries(libsirius ${SIRIUS_LINK_LIBS})
-target_link_libraries(libsirius-static ${SIRIUS_LINK_LIBS})
+target_link_libraries(libsirius ${LIBSIRIUS_LINK_LIBS})
+target_link_libraries(libsirius-static ${LIBSIRIUS_LINK_LIBS})
 
-install(TARGETS libsirius DESTINATION ${CMAKE_INSTALL_PREFIX}/lib)
-install(TARGETS libsirius-static DESTINATION ${CMAKE_INSTALL_PREFIX}/lib)
+install(TARGETS libsirius DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/sirius)
+install(TARGETS libsirius-static DESTINATION ${CMAKE_INSTALL_PREFIX}/lib/sirius)
 # copy sirius include directory into install directory
 install(FILES ${SIRIUS_CONFIG_OUT}
-        DESTINATION ${CMAKE_INSTALL_PREFIX}/include/sirius)
+        DESTINATION ${CMAKE_INSTALL_PREFIX}/include/sirius/sirius)
 install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/sirius
-        DESTINATION ${CMAKE_INSTALL_PREFIX}/include
+        DESTINATION ${CMAKE_INSTALL_PREFIX}/include/sirius
         FILES_MATCHING PATTERN "*.h" PATTERN "*.txx")
 # copy spdlog library into install directory
 install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/../third_party/spdlog/spdlog/include/spdlog
-        DESTINATION ${CMAKE_INSTALL_PREFIX}/include)
+        DESTINATION ${CMAKE_INSTALL_PREFIX}/include/sirius)
+# copy gsl library into install directory
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/../third_party/gsl/gsl/include/gsl
+        DESTINATION ${CMAKE_INSTALL_PREFIX}/include/sirius)
 
+if (ENABLE_SIRIUS_EXECUTABLE OR ENABLE_UNIT_TESTS)
+    set(LIBSIRIUS_GDAL_SRC
+        # gdal
+        sirius/gdal/debug.h
+        sirius/gdal/debug.cc
+        sirius/gdal/error_code.h
+        sirius/gdal/error_code.cc
+        sirius/gdal/exception.h
+        sirius/gdal/exception.cc
+        sirius/gdal/stream_block.h
+        sirius/gdal/input_stream.h
+        sirius/gdal/input_stream.cc
+        sirius/gdal/resampled_output_stream.h
+        sirius/gdal/resampled_output_stream.cc
+        sirius/gdal/types.h
+        sirius/gdal/wrapper.h
+        sirius/gdal/wrapper.cc)
+    add_library(libsirius-gdal-static STATIC EXCLUDE_FROM_ALL ${LIBSIRIUS_GDAL_SRC})
+    target_link_libraries(libsirius-gdal-static gdal libsirius)
+endif ()
 
-add_executable(sirius main.cc)
-target_link_libraries(sirius libsirius-static cxxopts Threads::Threads)
+if (ENABLE_SIRIUS_EXECUTABLE)
+    set(SIRIUS_MAIN_SRC
+        # main
+        sirius/image_streamer.h
+        sirius/image_streamer.cc
 
-install(TARGETS sirius DESTINATION ${CMAKE_INSTALL_PREFIX}/bin)
+        main.cc)
+
+    add_executable(sirius ${SIRIUS_MAIN_SRC})
+    target_link_libraries(sirius
+        libsirius-static
+        libsirius-gdal-static
+        cxxopts
+        Threads::Threads)
+
+    install(TARGETS sirius DESTINATION ${CMAKE_INSTALL_PREFIX}/bin)
+endif()
