+++ tests/filter_tests.cc
@@ -26,6 +26,7 @@
 #include "sirius/types.h"
 
 #include "sirius/fftw/wrapper.h"
+#include "sirius/gdal/wrapper.h"
 
 #include "sirius/utils/log.h"
 
@@ -47,8 +48,8 @@ TEST_CASE("filter - dirac filter", "[sirius]") {
     auto complex_array = sirius::fftw::CreateComplex(size);
     std::memset(complex_array.get(), 1, size.CellCount());
 
-    auto filter =
-          sirius::Filter::Create("./filters/dirac_filter.tiff", zoom_ratio);
+    auto filter = sirius::Filter::Create(
+          sirius::gdal::LoadImage("./filters/dirac_filter.tiff"), zoom_ratio);
     REQUIRE(filter.IsLoaded());
 
     auto output = filter.Process(size, std::move(complex_array));
@@ -57,6 +58,9 @@ TEST_CASE("filter - dirac filter", "[sirius]") {
 TEST_CASE("filter - sinc2 filter", "[sirius]") {
     LOG_SET_LEVEL(trace);
     auto zoom_ratio = sirius::ZoomRatio::Create(2, 1);
+    auto filter = sirius::Filter::Create(
+          sirius::gdal::LoadImage("./filters/sinc_zoom2_filter.tif"),
+          zoom_ratio);
     sirius::Size regular_size = {100, 100};
     sirius::Size large_size = {1000, 1000};
 
@@ -64,8 +68,6 @@ TEST_CASE("filter - sinc2 filter", "[sirius]") {
         sirius::Size size{10, 10};
         auto complex_array = sirius::fftw::CreateComplex(size);
         std::memset(complex_array.get(), 1, size.CellCount());
-        auto filter = sirius::Filter::Create("./filters/sinc_zoom2_filter.tif",
-                                             zoom_ratio);
         REQUIRE(filter.IsLoaded());
         REQUIRE_THROWS_AS(filter.Process(size, std::move(complex_array)),
                           sirius::Exception);
@@ -74,8 +76,6 @@ TEST_CASE("filter - sinc2 filter", "[sirius]") {
         sirius::Size size{100, 100};
         auto complex_array = sirius::fftw::CreateComplex(size);
         std::memset(complex_array.get(), 1, size.CellCount());
-        auto filter = sirius::Filter::Create("./filters/sinc_zoom2_filter.tif",
-                                             zoom_ratio);
         REQUIRE(filter.IsLoaded());
         REQUIRE_NOTHROW(complex_array =
                               filter.Process(size, std::move(complex_array)));
@@ -85,16 +85,12 @@ TEST_CASE("filter - sinc2 filter", "[sirius]") {
         sirius::Size size{1000, 1000};
         auto complex_array = sirius::fftw::CreateComplex(size);
         std::memset(complex_array.get(), 1, size.CellCount());
-        auto filter = sirius::Filter::Create("./filters/sinc_zoom2_filter.tif",
-                                             zoom_ratio);
         REQUIRE(filter.IsLoaded());
         REQUIRE_NOTHROW(complex_array =
                               filter.Process(size, std::move(complex_array)));
         REQUIRE_NOTHROW(filter.Process(size, std::move(complex_array)));
     }
     SECTION("sinc2 filter - image size is filter size") {
-        auto filter = sirius::Filter::Create("./filters/sinc_zoom2_filter.tif",
-                                             zoom_ratio);
         auto size = filter.size();
         auto complex_array = sirius::fftw::CreateComplex(size);
         std::memset(complex_array.get(), 1, size.CellCount());
